All components in a package
